compatible for wasm build
