+++ barretenberg/cpp/src/barretenberg/honk/execution_trace/execution_trace_block.hpp
@@ -12,6 +12,7 @@
 #include "barretenberg/common/ref_vector.hpp"
 #include "barretenberg/common/serialize.hpp"
 #include "barretenberg/common/slab_allocator.hpp"
+#include "barretenberg/common/slab_vector.hpp"
 #include "barretenberg/common/throw_or_abort.hpp"
 #include <cstddef>
 #include <cstdint>
@@ -227,7 +228,7 @@ template <typename FF, size_t NUM_WIRES_> class ExecutionTraceBlock {
     static constexpr size_t NUM_WIRES = NUM_WIRES_;
 
     using SelectorType = Selector<FF>;
-    using WireType = FileBackedVector<uint32_t>;
+    using WireType = SlabVector<uint32_t>;
     using Wires = std::array<WireType, NUM_WIRES>;
 
     ExecutionTraceBlock() = default;
